Move serial to pkg/ directory
