+++ cmd/tether/tether_linux.go
@@ -34,8 +34,8 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/kr/pty"
-	"github.com/vmware/vic/cmd/tether/serial"
 	"github.com/vmware/vic/pkg/dio"
+	"github.com/vmware/vic/pkg/serial"
 )
 
 // allow us to pick up some of the osops implementations when mocking
