+++ pkg/serial/rawconn.go
